+++ src/main/java/com/project/hospitalReport/dao/DoctorDao.java
@@ -3,7 +3,7 @@ package com.project.hospitalReport.dao;
 import java.time.LocalTime;
 import java.util.List;
 
-import com.project.hospitalReport.dto.*;
+import com.project.hospitalReport.entity.*;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import org.hibernate.Session;
@@ -11,7 +11,7 @@ import org.hibernate.Transaction;
 import org.hibernate.query.NativeQuery;
 import org.springframework.stereotype.Repository;
 
-import com.project.hospitalReport.helper.ConfigClass;
+import com.project.hospitalReport.configuration.ConfigClass;
 
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
@@ -196,7 +196,7 @@ public class DoctorDao {
 
 	public String addAppointment(Appointment a) {
 		Appointment appointment = new Appointment();
-		Integer pid = a.getId();
+		Long pid = a.getId();
 		Session session = ConfigClass.getSession().openSession();
 		Transaction transaction = session.beginTransaction();
 		try{
@@ -273,7 +273,7 @@ public class DoctorDao {
 	}
 
 	public String updateDrug(DrugsStock updated) {
-		Integer id = updated.getId();
+		Long id = updated.getId();
 		Session session = ConfigClass.getSession().openSession();
 		Transaction transaction = session.beginTransaction();
 		DrugsStock actual = session.get(DrugsStock.class, id);
@@ -321,7 +321,7 @@ public class DoctorDao {
 		return result;
 	}
 
-	public List<Object[]> getDrugById(Integer id){
+	public List<Object[]> getDrugById(Long id){
 		Session session = ConfigClass.getSession().openSession();
 		session.beginTransaction();
 		String sql = "SELECT ds.id, ds.name, ds.mrp, ds.perPieceRate, ds.addedDate, ds.quantity " +
